refactor: split api calls
