+++ lib/api/utils.ts
@@ -1,0 +1,4 @@
+export type APIConfig = {
+  baseURL: string;
+  apiKey: string;
+};
